moving the command specs to their own files
